Separação de responsabilidades e Mudança da UI
Criação de pastas com separação de responsabilidades e mudança da interface gráfica para tom black.
